modern: replace cast-as-lvalue assignments
Fix all 11 cast-as-lvalue occurrences in the four TUs flagged by the
modern GCC 13 sweep (GCC rejects the legacy GNU C extension of casting
an lvalue and assigning through it):

- src/banim-efxmagic-crimsoneye.c: two `(struct Anim *)(proc->unkNN) =
  ...` stores into int-typed pointer-aliasing fields become plain
  `proc->unkNN = (int)(...)` assignments — bit-identical on this
  32-bit target since int and pointer are the same width, and the
  fields are already read back via `(struct Anim *)proc->unkNN`
  elsewhere.
- src/banim-efxmagic-shine.c: `(s16) proc->unk30 += 2;` becomes
  `proc->unk30 += 2;` — unk30 is u16, and addition mod 2^16 yields the
  same bit pattern whether performed through the signed or unsigned
  view, so the cast added no behavior, only the invalid lvalue.
- src/eventcall.c and src/eventfx-earthquake.c: eight
  `(u16)gBmSt.camera.{x,y} &= MASK;` compound bitwise-AND assignments
  become direct `gBmSt.camera.{x,y} &= MASK;` — camera.x/y are signed
  short, but AND-with-mask followed by truncation back to 16 bits is
  unaffected by sign extension of the discarded upper bits, so the
  cast was redundant for this operation.

Verified with the modern toolchain (arm-none-eabi-gcc 13.2.1) across
all debug/release x aapcs/apcs-gnu cells for -fsyntax-only and -c: all
32 targeted cells (4 files x 2 configs x 2 ABIs x {syntax,object}) now
pass. A full 434-file src/ sweep across the same 8 cells shows syntax
pass 354->358 and object pass 352->356 (+4/+4), with the exact
fail->pass transition set equal to these four files and zero
regressions elsewhere. scripts/modernize/tests and
'audit.py --check' both pass unchanged.

The legacy agbcc/asmdiff ROM build could not be exercised in this
worktree: tools/agbcc is not built and baserom.gba is not present, so
legacy build/compare validation is left to parent integration.

Copilot-Session: 79809916-31bb-45e7-bb66-2a603a0c404f

+++ src/eventcall.c
@@ -255,7 +255,7 @@ void EventQuakefxHorizon_SlightLoop(struct Proc * proc)
     {
         if (GetGameClock() % 2)
         {
-            (u16)gBmSt.camera.x &= ~2;
+            gBmSt.camera.x &= ~2;
             gBmSt.camera.x ^= 1;
         }
     }
@@ -279,7 +279,7 @@ void EventQuakefxVeritical_Loop(struct Proc * proc)
     {
         if (GetGameClock() % 2)
         {
-            (u16)gBmSt.camera.y &= 0xFFFD;
+            gBmSt.camera.y &= 0xFFFD;
             gBmSt.camera.y ^= 1;
         }
     }
@@ -339,14 +339,14 @@ void StartEventHorizontalQuakefxSlightlyNoSound(ProcPtr parent)
 
 void EndEventHorizontalQuakefx(ProcPtr parent)
 {
-    (u16)gBmSt.camera.x &= 0xFFFC;
+    gBmSt.camera.x &= 0xFFFC;
     Proc_EndEach(ProcScr_EventHorizontalQuakefx);
     Sound_FadeOutSE(4);
 }
 
 void EndEventVerticalQuakefx(void)
 {
-    (u16)gBmSt.camera.y &= 0xFFFC;
+    gBmSt.camera.y &= 0xFFFC;
     Proc_EndEach(ProcScr_EventVerticalQuakefx);
     Sound_FadeOutSE(4);
 }
@@ -371,7 +371,7 @@ void EventQuakefx_Loop(struct Proc * proc)
     {
         if (GetGameClock() % 2)
         {
-            (u16)gBmSt.camera.x &= 0xFFFD;
+            gBmSt.camera.x &= 0xFFFD;
             gBmSt.camera.x ^= 1;
         }
     }
@@ -396,7 +396,7 @@ void StartEventQuakefx(ProcPtr proc)
 
 void EndEventQuakefx(ProcPtr proc)
 {
-    (u16)gBmSt.camera.y &= 0xFFFC;
+    gBmSt.camera.y &= 0xFFFC;
     Sound_FadeOutSE(4);
     Proc_EndEach(ProcScr_EventQuakefx);
 }
